virts-510-adding pwsh to path, added powershell core (pwsh) default (#165)
installation path to $env:Path in powershell to allow agent to spawn a
beacon in pwsh independent of powershell. Also added a -Wait flag to the
installer to ensure that pwsh is installed. After execution of pwsh,
powershell processes can be terminated without losing agent.
allow the agent to spawn pwsh

Former-commit-id: 581b5c518dff37607e4314b68f518a0df5004040